\name{klein14}
\docType{data}
\alias{klein14}

\title{MCMC results in Klein (2014), Table 8.}

\description{
   MCMC results in Klein (2014), Table 8.
}

\usage{data(baac00)}

\format{
   This list contains the following elements:
   \describe{
      \item{model.list}{.}
      \item{coefs}{.}
   }
}

\references{
Klein, T. (2014). Stable matching in microcredit: Implications for market design & econometric analysis, PhD thesis, \emph{University of Cambridge}.
}

\keyword{datasets}
