\name{klein15a}
\docType{data}
\alias{klein15a}

\title{MCMC results in Klein (2015a).}

\description{
   MCMC results in Klein (2015a).
}

\usage{data(klein15a)}

\format{
   This list contains the following elements:
   \describe{
      \item{model.list}{.}
      \item{coefs}{.}
   }
}

\references{
Klein, T. (2015a). Stable matching in microcredit: Implications for market design & econometric analysis, PhD thesis, \emph{University of Cambridge}.
}

\keyword{datasets}
