\name{mce15}
\docType{data}
\alias{mce15}

\title{Results of Monte Carlo Simulations in Klein (2015).}

\description{
   Results Monte Carlo Simulations in Klein (2015).
}

\usage{data(mce15)}

\format{
   This list contains the following elements:
   \describe{
      \item{exp.5.5.ols}{.}
      \item{exp.5.5.ntu}{.}
      \item{exp.6.5.ols}{.}
      \item{exp.6.5.ntu}{.}
      \item{exp.7.5.ols}{.}
      \item{exp.7.5.ntu}{.}
      \item{exp.6.6.ols}{.}
      \item{exp.6.6.ntu}{.}
      \item{exp.7.7.ols}{.}
      \item{exp.7.7.ntu}{.}
   }
}

\references{
Klein, T. (2014). Stable matching in microcredit: Implications for market design & econometric analysis, PhD thesis, \emph{University of Cambridge}.
}

\keyword{datasets}
