% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mfx}
\alias{mfx}
\title{Marginal effects for probit and matching models}
\usage{
mfx(m, toLatex = FALSE)
}
\arguments{
\item{m}{an object returned from function \code{stabit}.}

\item{toLatex}{logical: if \code{TRUE} the result tables are printed in Latex format. The default setting is \code{FALSE}.}
}
\description{
Marginal effects from regression coefficients for probit
and matching models.
}
\examples{
## 1. Read model results from Klein (2014), Table 5
## filepath <- system.file("scripts/TK_gibbsiter_ntu.mat", package="matchingMarkets")

## 2. Apply mfx function and print results
## mfx(m=filepath)
}
\author{
Thilo Klein
}
\keyword{summary}

