% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabchk2.R
\name{stabchk2}
\alias{stabchk2}
\title{Stability-Check}
\usage{
stabchk2(
  matching,
  c.prefs,
  s.prefs,
  nSlots = NULL,
  withTies = FALSE,
  envyEmpty = TRUE
)
}
\arguments{
\item{matching}{matrix (or data frame) of matchings. Data frame or matrix with two columns containing in column 1 the colleges and in column 2 the students with each row forming a couple.}

\item{c.prefs}{college preferences. If \code{withTies} is set to \code{FALSE}, matrix with \code{nColleges} columns. Otherwise, matrix of dimension \code{nStudents} x \code{nColleges}.}

\item{s.prefs}{student preferences. If \code{withTies} is set to \code{FALSE}, matrix with \code{nStudents} columns. Otherwise, matrix of dimension \code{nColleges} x \code{nStudents}.}

\item{nSlots}{capacity for each college. \code{NULL} or a vector of length \code{nColleges} giving the capacity for each college.}

\item{withTies}{logical. If \code{FALSE} (the default), ranks are inferred from the ordering of the preference matrices. Otherwise, ranks must be explicitly given in the preference matrices (see Details).}

\item{envyEmpty}{logical. Only used if \code{nSlots} is \code{NULL}. If \code{TRUE} (the default), students may have justified envy towards empty colleges (see Details).}
}
\value{
\code{stabchk} returns a data frame with as many rows as blocking pairs were found. Column 1 indicates the college and column 2 indicate the student of the blocking pairs.
}
\description{
Checks a given two sided matching for blocking pairs.
}
\details{
If \code{withTies} is set to \code{FALSE}, then the default format for the preference matrices is expected, i.e. \code{c.prefs[r,c]} contains the student of rank \code{r} in the ranking of college \code{c}. Missing values are used to indicate unacceptable students/colleges. Especially, this does not allow for ties in the rankings, since ranks are inferred by row. If \code{withTies} is set to \code{TRUE}, then the preference matrices must explicitly contain the ranks, i.e. \code{c.prefs[s,c]} contains the rank of student \code{s} in the ranking of college \code{c}.

If \code{nSlots} is not \code{NULL}, then any college with fewer assigned students than available slots is treated as if it was assigned an unacceptable student. Otherwise, this is controlled by \code{envyEmpty}. If \code{envyEmpty} is set to \code{TRUE} (the default), then any college without any matches is treated as if it was assigned an unacceptable student, i.e. it forms a blocking pair with any acceptable student that prefers this college over its assignment. Otherwise, such a college cannot be part of any blocking pair.
}
