\name{klein15a}
\docType{data}
\alias{klein15a}

\title{MCMC results in Klein (2015a)}

\description{
   MCMC results in Klein (2015a).
}

\usage{data(klein15a)}

\format{
   A list containing the following elements:
   \describe{
      \item{model.list}{.}
      \item{coefs}{.}
   }
}

\references{
Klein, T. (2015a). \href{https://ideas.repec.org/p/cam/camdae/1521.html}{Does Anti-Diversification Pay? A One-Sided Matching Model of Microcredit}. \emph{Cambridge Working Papers in Economics}, #1521.
}

\keyword{datasets}
