% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mfx}
\alias{mfx}
\title{Marginal effects for probit and matching models}
\usage{
mfx(m, toLatex = FALSE)
}
\arguments{
\item{m}{an object returned from function \code{stabit}.}

\item{toLatex}{logical: if \code{TRUE} the result tables are printed in Latex format. The default setting is \code{FALSE}.}
}
\description{
Marginal effects from regression coefficients for probit
and matching models.
}
\examples{
## 1. load results from Klein (2014), Table 8
 data(klein14)

## 2. apply mfx function and print results
 mfx(m=klein14)
}
\author{
Thilo Klein
}
\keyword{summary}

